backend/requirements.txt
django
djangorestframework
django-filter
pillow
djangorestframework-simplejwt
psycopg2
psycopg2-binary
python-dotenv
requests
geoserver-restconfig